few updates

+++ src/app/bill/components/bill-list/bill-list.component.ts
@@ -1,9 +1,10 @@
-import { Component, Input, OnInit, SimpleChanges } from '@angular/core';
+import { Component, DestroyRef, Input, OnInit, SimpleChanges, ViewChild } from '@angular/core';
 import { MatTableDataSource } from '@angular/material/table';
-import { IBIllDto } from 'src/app/core/models';
+import { IBIllDto, ISubCategory } from 'src/app/core/models';
 import { BillService } from '../../services/bill.service';
 import { Router } from '@angular/router';
 import { AlertIcon, AlertService } from 'src/app/shared/services/alert.service';
+import { MatPaginator } from '@angular/material/paginator';
 
 @Component({
   selector: 'app-bill-list',
@@ -13,13 +14,19 @@ import { AlertIcon, AlertService } from 'src/app/shared/services/alert.service';
 export class BillListComponent implements OnInit {
   @Input() bills: IBIllDto[] = [];
 
-  displayedColumns: string[] = ['id', 'payment', 'total', 'totalIncoming', 'actions'];
+  displayedColumns: string[] = ['payment', 'total', 'totalIncoming', 'actions'];
 
   dataSource = new MatTableDataSource<IBIllDto>();
+
+  @ViewChild(MatPaginator) paginator: MatPaginator | undefined;
 
   constructor(private billService: BillService,
     private router: Router,
     private alertService: AlertService) { }
+
+  ngAfterViewInit(): void {
+    this.dataSource.paginator = this.paginator as MatPaginator;
+  }
 
   ngOnInit(): void {
     this.dataSource.data = this.bills;
@@ -42,7 +49,7 @@ export class BillListComponent implements OnInit {
       icon: AlertIcon.Warning,
     });
 
-    if (confirmRemove) {
+    if (confirmRemove && bill.id) {
       this.billService.delete(bill.id)
         .subscribe(() => {
           this.alertService.toastAlert({
